# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Waryaa Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, username, rank, tournamentWins, winRate, matchesPlayed, matchesWon, points, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        rank:
          type: integer
        tournamentWins:
          type: integer
        winRate:
          type: number
        matchesPlayed:
          type: integer
        matchesWon:
          type: integer
        points:
          type: integer
        country:
          type: ["string", "null"]
        discordId:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isActive:
          type: boolean
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Team:
      type: object
      required: [id, name, wins, losses, points, memberCount]
      properties:
        id:
          type: integer
        name:
          type: string
        tag:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        captainId:
          type: integer
        captainName:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        points:
          type: integer
        memberCount:
          type: integer
        description:
          type: ["string", "null"]
        achievements:
          type: array
          items:
            type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        createdAt:
          type: string
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        tag:
          type: string
        description:
          type: string
...